+++ polar-core/src/lib.rs
@@ -9,16 +9,15 @@ mod lexer;
 #[macro_use]
 pub mod macros;
 pub mod events;
-mod kb;
-mod messages;
+pub mod kb;
+pub mod messages;
 mod numerics;
 pub mod parser;
 pub mod polar;
 mod rewrites;
-mod rules;
+pub mod rules;
 mod sources;
 pub mod terms;
 pub mod traces;
-pub mod types;
 mod vm;
 mod warnings;
