+++ polar-core/tests/serialize.rs
@@ -1,0 +1,72 @@
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+
+    use polar_core::{error::*, events::*, rules::*, term, terms::*, value};
+
+    #[test]
+    fn serialize_test() {
+        let pred = Predicate {
+            name: Symbol("foo".to_owned()),
+            args: vec![Term::new_from_test(value!(0))],
+        };
+        assert_eq!(
+            serde_json::to_string(&pred).unwrap(),
+            r#"{"name":"foo","args":[{"value":{"Number":{"Integer":0}}}]}"#
+        );
+        let event = QueryEvent::ExternalCall {
+            call_id: 2,
+            instance: None,
+            attribute: Symbol::new("foo"),
+            args: Some(vec![
+                Term::new_from_test(value!(0)),
+                Term::new_from_test(value!("hello")),
+            ]),
+        };
+        eprintln!("{}", serde_json::to_string(&event).unwrap());
+        let term = Term::new_from_test(value!(1));
+        eprintln!("{}", serde_json::to_string(&term).unwrap());
+        let mut fields = BTreeMap::new();
+        fields.insert(Symbol::new("hello"), term!(1234));
+        fields.insert(
+            Symbol::new("world"),
+            Term::new_from_test(Value::String("something".to_owned())),
+        );
+        let literal = InstanceLiteral {
+            tag: Symbol::new("Foo"),
+            fields: Dictionary { fields },
+        };
+        let event = QueryEvent::MakeExternal {
+            instance_id: 12345,
+            constructor: Term::new_from_test(Value::InstanceLiteral(literal)),
+        };
+        eprintln!("{}", serde_json::to_string(&event).unwrap());
+        let external = Term::new_from_test(Value::ExternalInstance(ExternalInstance {
+            instance_id: 12345,
+            constructor: None,
+            repr: None,
+        }));
+        let list_of = Term::new_from_test(Value::List(vec![external]));
+        eprintln!("{}", serde_json::to_string(&list_of).unwrap());
+        let mut fields = BTreeMap::new();
+        fields.insert(Symbol::new("foo"), list_of);
+        let dict = Term::new_from_test(Value::Dictionary(Dictionary { fields }));
+        eprintln!("{}", serde_json::to_string(&dict).unwrap());
+        let e = error::ParseError::InvalidTokenCharacter {
+            token: "Integer".to_owned(),
+            c: 'x',
+            loc: 99,
+        };
+        let err: PolarError = e.into();
+        eprintln!("{}", serde_json::to_string(&err).unwrap());
+        let rule = Rule {
+            name: Symbol::new("foo"),
+            params: vec![],
+            body: Term::new_temporary(Value::Expression(Operation {
+                operator: Operator::And,
+                args: vec![dict.clone(), dict.clone(), dict],
+            })),
+        };
+        eprintln!("{}", rule);
+    }
+}
